move monaco initialization, theme and editor support instance out of monaco component
