+++ src/browser/modules/Stream/EditFrame.tsx
@@ -25,25 +25,20 @@ import { Action } from 'redux'
 import styled from 'styled-components'
 import { Bus } from 'suber'
 
-import { BrowserTheme } from '../Editor/CypherMonacoThemes'
 import Monaco from '../Editor/Monaco'
 import FrameTemplate from '../Frame/FrameTemplate'
 import { StyledFrameBody, StyledFrameContents } from '../Frame/styled'
-import useDerivedTheme from 'browser-hooks/useDerivedTheme'
 import {
   commandSources,
   executeCommand
 } from 'shared/modules/commands/commandsDuck'
 import {
   codeFontLigatures,
-  getTheme,
-  LIGHT_THEME,
   shouldEnableMultiStatementMode
 } from 'shared/modules/settings/settingsDuck'
 import { Frame, GlobalState } from 'shared/modules/stream/streamDuck'
 
 interface EditFrameProps {
-  browserTheme: BrowserTheme
   bus: Bus
   codeFontLigatures: boolean
   enableMultiStatementMode: boolean
@@ -64,10 +59,6 @@ const ForceFullSizeFrameContent = styled.div`
 const EditFrame = (props: EditFrameProps): JSX.Element => {
   const [text, setText] = useState(props.frame.query)
 
-  const [theme] = useDerivedTheme(props.browserTheme, LIGHT_THEME) as [
-    BrowserTheme
-  ]
-
   return (
     <ForceFullSizeFrameContent>
       <FrameTemplate
@@ -78,7 +69,6 @@ const EditFrame = (props: EditFrameProps): JSX.Element => {
             fontLigatures={props.codeFontLigatures}
             id={props.frame.id}
             onChange={setText}
-            theme={theme}
             value={text}
           />
         }
@@ -92,7 +82,6 @@ const EditFrame = (props: EditFrameProps): JSX.Element => {
 }
 
 const mapStateToProps = (state: GlobalState) => ({
-  browserTheme: getTheme(state),
   enableMultiStatementMode: shouldEnableMultiStatementMode(state),
   codeFontLigatures: codeFontLigatures(state)
 })
